Fix parse_block_html settings
Now using inline parse block html settings.

This setting enables Kramdown to parse markdown within HTML tags on
markdown files. Previously it was configured to work across the site.
It is now enabled only when explicitly stated.

Typically used for the Table of Contents and also for collapsible
code cells such as on the Central Limit Theorem article.